♻ use one example list for everything
